configuration in entity id of long

+++ SpringBoot/place/src/main/java/com/market/place/domain/repository/UserRepository.java
@@ -9,10 +9,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserRepository {
-    Optional<UserDomain> getUser(int userId);
+    Optional<UserDomain> getUser(Long userId);
     List<UserDomain> getAll();
     Optional<List<UserDomain>> getAllUsersActive(int active);
     UserDomain save(UserDomain userDomain);
     Optional<Page<UserEntity>> getPagination(Pageable pageable);
-    void delete(int userId);
+    void delete(Long userId);
 }
