+++ SpringBoot/place/src/main/java/com/market/place/domain/service/AcountService.java
@@ -24,10 +24,10 @@ public class AcountService {
     }
     public UserDomain save(UserDomain userDomain){return userRepository.save(userDomain);}
     public Optional<Page<UserEntity>> getPagination(Pageable pageable){ return userRepository.getPagination(pageable);}
-    public Optional<UserDomain> getUser(int userId) {
+    public Optional<UserDomain> getUser(Long userId) {
         return userRepository.getUser(userId);
     }
-    public boolean delete(int userId) {
+    public boolean delete(Long userId) {
         return getUser(userId).map(user -> {
             userRepository.delete(userId);
             return true;
