+++ SpringBoot/place/src/main/java/com/market/place/persistence/crud/UserCrudRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserCrudRepository extends CrudRepository<UserEntity, Integer> {
+public interface UserCrudRepository extends CrudRepository<UserEntity, Long> {
     Optional<List<UserEntity>> findByActive(int active);
 }
